add bolg request model.
