+++ BlogSystem.Business/.gitignore
@@ -1,3 +1,4 @@
 .vscode/
 bin/
 obj/
+.DS_Store
