+++ BlogSystem.Business/Domain/Blog.cs
@@ -1,25 +1,15 @@
 using BlogSystem.DataAccess;
+using System.ComponentModel.DataAnnotations;
 
 namespace BlogSystem.Business
 {
     public class Blog
     {
         public long Id { get; set; }
+
+        [Required]
         public string Title { get; set; }
+
         public string Content { get; set; }
-
-        // ? Should it be a extension method?
-        public BlogEntity ToBlogEntity()
-        {
-            if (this == null)
-                return null;
-
-            return new BlogEntity
-            {
-                Id = this.Id,
-                Title = this.Title,
-                Content = this.Content
-            };
-        }
     }
 }
